Data Ingestion Module/requirements.txt
fastapi==0.95.2
pydantic==2.0.3
pandas==2.0.2
numpy==1.25.0
pyarrow==11.0.0
kafka-python==3.0.1
psycopg2-binary==2.9.6
requests==2.30.0
uvicorn==0.22.0
pytest==7.4.0

# Install dependencies with: pip install -r requirements.txt
